requirements.txt
azure-functions
openai==0.28.0
azure-identity
azure-mgmt-compute
azure-mgmt-storage
